Moved menu to new page + introduced specialities section as menu summary
